requirements.txt
ccxt>=4.0.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
plotly>=5.3.0
python-dotenv>=0.19.0
pytest>=6.2.5
aiohttp>=3.8.0
python-dateutil>=2.8.2
dash>=2.0.0
dash-bootstrap-components>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0